refactor: fetch Firebase data when each page mounted
